feat: 补充标签接口

+++ src-tauri/src/db/mod.rs
@@ -60,6 +60,6 @@ pub struct PageParams {
 /// Common return Object
 #[derive(Debug, Deserialize, Serialize)]
 pub struct PageResult<T> {
-    pub total: usize,
+    pub total: u64,
     pub rows: Vec<T>,
 }
